feat: add schema migration feature and sequential examples
- Implement Migration struct and RunMigrations function in migration.go
- Add tracking of applied migrations via schema_migrations table
- Add unit tests for successful migrations and rollbacks
- Provide sequential examples in examples/01_init and examples/02_migrate
